Refactored main nav into smaller, more focused components
