+++ react/new/src/constants/navigation.ts
@@ -1,0 +1,17 @@
+export const NAVIGATION_CONSTANTS = {
+  MENU_Z_INDEX: 300,
+
+  // Animation timings
+  MENU_TRANSITION_DURATION: 200,
+  MOBILE_MENU_TRANSITION_DURATION: 300,
+
+  // Layout dimensions
+  HEADER_HEIGHT: 60,
+  LOGO_HEIGHT: 28,
+
+  // Breakpoints
+  MOBILE_BREAKPOINT: 768,
+
+  // Menu behavior
+  DROPDOWN_EXIT_DURATION_RATIO: 0.5, // For exitDuration calculation
+} as const
